refactor(ui): split ZoneRecipesSection into per-component modules (#456, #387)
ZoneRecipesSection.tsx was 2201 lines holding ~22 components in one file.
Extract them into focused modules under ui/src/components/recipes/, bottom-up
so no import cycles form and each cut left the tree compiling:

- recipe-slot-helpers.ts  — SlotChunk + groupSlots/isGroupFilled/isGroupRequired/
                            getGroupKeys + durationToMinutes (pure helpers)
- recipe-form-fields.tsx  — leaf inputs (EquipmentOptions, MultiSelectChips,
                            DurationInput, TimeInput, SingleEquipmentZonePicker,
                            EquipmentListPicker, EquipmentCheckboxList) plus
                            ModeCyclePill and CountdownTimer
- RecipePickerPopover.tsx, RecipeInstanceRow.tsx, DuplicateRecipeModal.tsx,
  AddRecipeForm.tsx
- ZoneRecipesSection.tsx now only composes the extracted pieces (120 lines)

#387 folded in: a shared useZoneOptions(zoneTree) hook returns { allZones,
zoneChains } and replaces the three duplicated flattenZonesWithPath/zoneChainMap
useMemo pairs in RecipeInstanceRow, DuplicateRecipeModal and AddRecipeForm.

Pure refactor: component bodies copied verbatim, no behavior change. tsc -b and
the full UI test suite (347 tests) stay green.
